Reorganized models.py for models folder with file for each type. Also removed intermediate models
